Added a temporary loading status for when the user is waiting on a return (not yet using Promises status)
